Added undo and redo
Added initial undo and redo, it does not know how to handle new input to rechange the order of the redo and undo. Left to implement: get rid of redos if new input is registered by the user. Added a struct node* previous to the node strust essentially it making it a doubly linked list behaving like a queue. Maybe not good? But works well *CHECK*