Finished Undo and redo functionality
removed bugs with undo when it would it take twice to press undo to undo. Added updateQueue method to update the queue accordingly if player undos and then plays a new move. It gets rid of all the "redo" moves and updates the queue. Bugs might appear, need more testing.